vip_app/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.1.0
pywebpush==2.0.3
gunicorn==23.0.0
psycopg[binary]==3.2.9
